apps/worker/pyproject.toml
[project]
name = "buildlaw-worker"
version = "0.1.0"
description = "BuildLaw AI background worker (Redis / RQ)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "redis>=5.2.0",
    "rq>=2.1.0",
    "psycopg2-binary>=2.9.10",
    "boto3>=1.35.0",
    "pypdf>=4.2.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
